backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.7.3
pydantic-settings==2.3.3
alembic==1.13.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2
apscheduler==3.10.4
pulp==2.8.0
prometheus-client==0.20.0
boto3==1.34.131
sentry-sdk==2.7.1
celery==5.4.0
redis==5.0.7
